Reorder raw message processing for variable rsize values
When processing the data stream coming from the PLM, we look for sent messages
first (with attached ack/nak) before trying to suss out any non-response
messages.  Since the rsize varies by code, we'll need to expand the code
database to include metadata about rsize so that commands can be successfully
pulled out of the FIFO receive buffer.

It's all still just bytes at this stage, but the plumbing is improving.