Rework buffer processing completely
After considerable experimentation it turns out that some of the assumptions
I made about the PLM behavior were incorrect which caused the old send logic to
cause problems.  In practice, if you've sent a command to the PLM then the only
valid thing to do at that point is to wait for the ack/nak for that command.
Until you've received the ack/nak you can't even begin to send another, stacked
command to the device.

As a result, I ripped out the "sent_messages" list and replaced it with just
a single "last_command" and I refuse to send more commands if this is non-NULL.

I also added a recv_queue globally to handle message processing from the
device and a send_queue to stack sent commands if we're blocked by
a last_command that's not ack/nak'd yet.