Include timeout when waiting for folloups
Sometimes the nature of the protocol means that we are expecting a response
first from the PLM and then a second, followup response from a referenced
device.  In these situations, the nature of the communications requires that we
wait until both messages have arrived.  However, if the actual device is dead
or missing, the second message will never arrive.

This commit adds a default timeout to the wait_for() functions which will give
up after [n] seconds (default 2) in the event that the message we're expecting
does not arrive.  It will then proceeed processing the send queue if the
timeout is reached.