Generic handle to compare message with criteria
I amended the existing wait_for processing to be a generic handler for testing
to see if a messagestring matches a criteria dict.  This will allow the same
code to be used for wait_for as well as callback evaluation, and ultimately
replace all the hard-coded if/elif branching for internal callbacks as well.