Support for alternate ALL-Link Record format
The USB Stick (2448A7) has a different message format for the ALL-Link
Record Response (0x57) where it sends the device cat/subcat/firmware in
a different order than we see from the older 2413 PLM devices.  So now
we store the PLM's cat/subcat we got from the initial connection and use
that to decide how to parse the ALDB messages when they arrive.