Created a base class for thermostats meant to handle multiple variants of the Insteon thermostat.  Currently they are identical, but they won't be shortly.
Fixed an apparent bug in states/thermostat where it referred to _update_method
when the instance member of state is actually named _updatemethod.  Note that
this really should either be included in the constructor (best choice) or
added as a proper property.  If the convention is that things starting with _
are private to the class, then external classes shouldn't be reading or
writing them.